feat(gateway): add Responses WebSocket probe profiles
