+++ apps/aether-gateway/src/ai_serving/planner/standard/openai/mod.rs
@@ -23,7 +23,7 @@ pub(crate) use responses::{
     build_local_openai_responses_stream_plan_and_reports_for_kind,
     build_local_openai_responses_sync_attempt_source_for_kind,
     build_local_openai_responses_sync_plan_and_reports_for_kind,
-    maybe_build_codex_responses_websocket_decision,
+    maybe_build_responses_websocket_decision,
     maybe_build_stream_local_openai_responses_decision_payload,
-    maybe_build_sync_local_openai_responses_decision_payload,
+    maybe_build_sync_local_openai_responses_decision_payload, ResponsesWebSocketDecision,
 };
