+++ apps/aether-gateway/src/orchestration/mod.rs
@@ -47,8 +47,8 @@ pub(crate) use self::policy::{
     append_local_failover_policy_to_value, codex_cyber_flag_passthrough_enabled,
     codex_responses_websocket_enabled, cyber_continue_failover_enabled,
     local_failover_policy_from_report_context, local_failover_policy_from_transport,
-    resolve_local_failover_policy, LocalFailoverPolicy, LocalFailoverRegexRule,
-    CYBER_CONTINUE_FAILOVER_CONFIG_KEY,
+    resolve_local_failover_policy, responses_websocket_adapter, LocalFailoverPolicy,
+    LocalFailoverRegexRule, ResponsesWebSocketAdapter, CYBER_CONTINUE_FAILOVER_CONFIG_KEY,
 };
 pub(crate) use self::recovery::{
     analyze_local_failover, apply_provider_failure_disposition, recover_local_failover_decision,
